+++ .github/workflows/checks.yml
@@ -43,10 +43,10 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Run Tests
-        run: poetry run nox -s lint
+        run: poetry run nox -s lint:code
 
       - name: Run type-check
-        run: poetry run nox -s type-check
+        run: poetry run nox -s lint:typing
 
       - name: Upload Artifacts
         uses: actions/upload-artifact@v3
